feat: 움직임만을 검증하는 PieceMovement 클래스 생성 및 검증 로직 구현
